Fixed deferred calls data base deletion.
git-svn-id: https://cvet.by:8443/svn/FOnlineSource/Main@1684 c2c6a35e-ad2b-4e1f-bb64-97b989be360b

Former-commit-id: d7ab09ab6af8883e9b9655733b776ccc4d34b15e

+++ Source/ScriptInvoker.cpp
@@ -114,11 +114,12 @@ bool ScriptInvoker::CancelDeferredCall( uint id )
     {
         if( it->Id == id )
         {
-            deferredCalls.erase( it );
-
             #ifdef FONLINE_SERVER
-            DbStorage->Delete( "DeferredCalls", id );
+            if( it->Saved )
+                DbStorage->Delete( "DeferredCalls", id );
             #endif
+
+            deferredCalls.erase( it );
             return true;
         }
     }
@@ -162,7 +163,8 @@ void ScriptInvoker::Process()
                 it = deferredCalls.erase( it );
 
                 #ifdef FONLINE_SERVER
-                DbStorage->Delete( "DeferredCalls", call.Id );
+                if( call.Saved )
+                    DbStorage->Delete( "DeferredCalls", call.Id );
                 #endif
 
                 RunDeferredCall( call );
